# Streaming store-gateway: make room for tenant ID in postings cache key