… CHANGELOG.md
… pkg/storegateway/storepb/cache.pb.go