@dimitarvdimitrov 2023-01-02
i also did a comparison of the size of entries in memcached before and after this change

on `main`: `(series + shard index + shard count) | protobufMarshal | snappy`

on this PR: `(series + (postings | diffEncode | snappy)) | protobufMarshal | snappy` 

these are the common benchmarks between the two. It looks like postings aren't adding almost anything. Could be because of the way they are generated in tests. Not sure if I should investigate

```
BenchmarkFetchCachedSeriesForPostings/6000_series_with_6_labels_each
    length_without_postings 47686 length_with_postings 47713
BenchmarkFetchCachedSeriesForPostings/1000_series_with_1_matcher
    length_without_postings 6108  length_with_postings 6121
BenchmarkFetchCachedSeriesForPostings/6000_series_with_6_labels_with_more_repetitions
    length_without_postings 48850 length_with_postings 48877
```