@dimitarvdimitrov 2023-01-03
i ran a quick test with this in our dev cluster. The test scenario was running many requests selecting ~50K series each without chunks. zone-a was running r218 with streaming and without mmap, zone-b was running with streaming and without mmap and with the changes in this PR, zone-c was running the default implementations

i didn't observe any negative impact on CPU or memory utilization, or latency

![Screenshot 2023-01-03 at 14 45 32](https://user-images.githubusercontent.com/21020035/210370047-ed44f24b-3a67-435c-9a66-c6056a7c05fd.png)
![Screenshot 2023-01-03 at 14 45 47](https://user-images.githubusercontent.com/21020035/210370055-0756b9f5-7d2a-4199-9bec-d36f748c1a1d.png)