@dimitarvdimitrov 2023-01-03
the store-gateway-load-test results weren't affected. And also SeriesForPostings cache hit rate was ~100% for zone-b (means that cache was actually used) 
![Screenshot 2023-01-03 at 16 47 10](https://user-images.githubusercontent.com/21020035/210391858-0a9e481b-a046-4f33-b19a-fb330fd190f9.png)